requirements.txt
python-dotenv
websockets